Refactor HetznerDNSProviderSolver into internal module
